refactor: extract GitHubExportProvider+WireTypes (#639 slice A)
First foundational slice of #639. Moves the 9 nested private wire-
type structs from `GitHubExportProvider.swift` (lines 736-821) into
a new `Sources/BugNarrator/Services/GitHubExportProvider+WireTypes.swift`
file, so subsequent slices B-G can reference them from separate
extension files.

Moved (byte-verbatim except for the widening keyword):
- GitHubIssueRequest
- GitHubIssueResponse (with nested CodingKeys enum)
- GitHubRepositoryValidationResponse (with nested CodingKeys enum)
- GitHubRepositoryListItem (with nested CodingKeys enum)
- GitHubRepositoryOwner
- GitHubRepositoryPermissions (with `canCreateIssues` computed var)
- GitHubSearchResponse
- GitHubSearchItem (with nested CodingKeys enum)
- GitHubErrorResponse

Visibility widening: each of the 9 structs widens from `private
struct` to (default-internal) `struct`. This is required because
the base `actor GitHubExportProvider` (in the original file) and
future extension files (planned slices B-G) both need to reference
these types across file boundaries. Nested `CodingKeys` enums
implicitly inherit the parent's widened visibility — no further
change required.

External impact: zero. `rg` across `Sources/` and `Tests/` confirms
the only production reference is intra-file (`GitHubExportProvider`
uses them internally); the only test reference is a doc comment.
`GitHubRepositoryPermissions.canCreateIssues` continues to be read
at the two existing call sites (:66, :192).

Line-count delta:
- GitHubExportProvider.swift: 821 -> 734 (-87 lines)
- GitHubExportProvider+WireTypes.swift (new): 88 lines
- xcodeproj: +4 (xcodegen)

Test coverage: `GitHubExportProviderTests` continues to pass
unchanged (its own test-only `GitHubIssueRequestPayload` type is
unaffected).

Closes #873. Continues #639 (design sketch on parent) after PR
#699 (#630 slice C RawTranscriptSection).
